fix: Rename project from django-admin-actions to django-admin-action-hero.
Thanks BBL!
