+++ docs/api/action_hero.actions.rst
@@ -1,0 +1,29 @@
+action\_hero.actions package
+============================
+
+Submodules
+----------
+
+action\_hero.actions.queue\_celery module
+-----------------------------------------
+
+.. automodule:: action_hero.actions.queue_celery
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+action\_hero.actions.simple module
+----------------------------------
+
+.. automodule:: action_hero.actions.simple
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: action_hero.actions
+   :members:
+   :show-inheritance:
+   :undoc-members:
