--- docs/api/admin_actions.actions.rst
@@ -1,29 +1,0 @@
-admin\_actions.actions package
-==============================
-
-Submodules
-----------
-
-admin\_actions.actions.queue\_celery module
--------------------------------------------
-
-.. automodule:: admin_actions.actions.queue_celery
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-admin\_actions.actions.simple module
-------------------------------------
-
-.. automodule:: admin_actions.actions.simple
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: admin_actions.actions
-   :members:
-   :show-inheritance:
-   :undoc-members:
